socNetwork/requirements.txt
djangorestframework
dj-config-url==0.1.1
dj-database-url==0.4.2
Django==1.11.5
django-bootstrap-form==3.3
django-bootstrap-toolkit==2.15.0
django-bootstrap3==9.1.0
django-widget-tweaks==1.4.1
gunicorn==19.7.1
mysqlclient==1.3.12
Pillow==3.1.2
PyMySQL==0.7.11
virtualenv==15.1.0
whitenoise==3.3.1
psycopg2==2.7.3.2
